+++ src/parsers/index.ts
@@ -5,18 +5,22 @@ import { EcmascriptParser } from './ecmascript'
 import { IniParser } from './ini'
 import { PoParser } from './po'
 import { PhpParser } from './php'
+import { Properties } from './properties'
 import { Parser } from './base'
 
 export const DefaultEnabledParsers = ['json', 'yaml', 'json5']
 
 export const AvaliablePasers: Parser[] = [
+  // enabled parsers
   new JsonParser(),
   new YamlParser(),
   new Json5Parser(),
 
+  // avaliable parsers
   new EcmascriptParser('js'),
   new EcmascriptParser('ts'),
   new IniParser(),
   new PoParser(),
   new PhpParser(),
+  new Properties(),
 ]